# Refactor row iteration in 'ImageRect::to_image`