--- a/AUTHORS
+++ b/AUTHORS
@@ -27,3 +27,4 @@ Inflation <2375962+inflation@users.noreply.github.com>
 Jacob Abel <jacobabel@nullpo.dev>
 Zoltan Szabadka
 Caio <c410.f3r@gmail.com>
+Louis Dispa <louis.dispa@outlook.fr>