@veluca93 2025-09-03
That's honestly a bit disappointing of the rust compiler...

The speedup certainly doesn't hurt, although I imagine that in the final version of the code to_image will be called extremely rarely if at all (in fact, I am not convinced that Image as it is now will continue existing).

Having said that, no harm in speeding things up in the meantime ;)